# modules/compliance/openapi.yaml
openapi: 3.1.0
info:
  title: DOS Compliance Module API
  description: |
    Compliance module HTTP API. Co-hosted by `governance-policy-service` via
    `@dos/module-compliance.registerCompliance()`. All routes are tenant-scoped
    via the JWT `tenant_id` claim and enforce per-resource permissions.

    Owns 9 tenant-scoped tables (`frameworks`, `controls`, `compliance_mappings`,
    `compliance_assessments`, `compliance_gaps`, `compliance_requirements`,
    `control_objectives`, `control_testing`, `control_evidence_links`),
    publishes 13 events, subscribes to 8.

    This spec covers the route surface declared in `module.manifest.json`
    `routeBases[]`. Per-endpoint Zod request/response schemas are added in
    Wave 2 (schema-validation sweep); per-endpoint detail is finalized for
    public publication in Wave 19.
  version: 1.0.0
  contact:
    name: product-shahin-ai
    email: compliance-eng@dos-aio.com
  license:
    name: Proprietary
    identifier: LicenseRef-DOS-AIO-Proprietary
servers:
  - url: https://shahin-ai.com
    description: Production
  - url: https://staging.shahin-ai.com
    description: Staging
  - url: http://localhost:4000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/compliance:
    get:
      x-zod-schemas:
        query: listQuerySchema
      tags: [overview]
      summary: Compliance overview / dashboard composite
      description: Returns posture summary, KPIs, top gaps, recent activity.
      responses:
        '200': { $ref: '#/components/responses/OverviewResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/compliance/controls/:id/test:
    post:
      summary: /api/compliance/controls/:id/test (POST)
      tags: [auto-generated]
      responses:
        '200': { description: OK }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/controls/tests:
    get:
      summary: /api/controls/tests (GET)
      tags: [auto-generated]
      responses:
        '200': { description: OK }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      summary: /api/controls/tests (POST)
      tags: [auto-generated]
      responses:
        '200': { description: OK }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/csa-responses/csa-responses:
    get:
      summary: /api/csa-responses/csa-responses (GET)
      tags: [auto-generated]
      responses:
        '200': { description: OK }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      summary: /api/csa-responses/csa-responses (POST)
      tags: [auto-generated]
      responses:
        '200': { description: OK }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/csa-responses/csa-responses/:id:
    get:
      summary: /api/csa-responses/csa-responses/:id (GET)
      tags: [auto-generated]
      responses:
        '200': { description: OK }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        DAuth-issued JWT with `tenant_id` + `sub` (user) + `permissions[]` claims.
        Issued by Keycloak (realm `dogan`); verified by every host service.
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              examples: [bad_input, not_found, forbidden, unauthorized, conflict, internal]
            message:
              type: string
            details:
              type: object
              description: Field-level validation errors when code=bad_input
            request_id:
              type: string
              description: Echoed for support correlation
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Permission denied (RBAC + ownership check)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    OverviewResponse:
      description: Compliance posture summary
      content:
        application/json:
          schema:
            type: object
            properties:
              postureScore: { type: number }
              frameworksTracked: { type: integer }
              controlsTotal: { type: integer }
              controlsEffective: { type: integer }
              gapsOpen: { type: integer }
              attestationsPending: { type: integer }
              recentActivity: { type: array, items: { type: object } }
...